+++ src/printer/sla/index.js
@@ -70,11 +70,6 @@ const sla = {
       module: updateHostname(dashboard),
     },
     {
-      path: "projects",
-      html: require("../../views/projects.html"),
-      module: updateHostname(projects),
-    },
-    {
       path: "temperature",
       html: require("../../views/temperature.html"),
       module: updateHostname(temperature),
@@ -96,6 +91,13 @@ const sla = {
       html: require("../../views/refill.html"),
       module: updateHostname(refill),
     },
+    process.env.WITH_PROJECTS ?
+      {
+        path: "projects",
+        html: require("../../views/projects.html"),
+        module: updateHostname(projects),
+      }
+      : null,
     process.env.WITH_SETTINGS ?
       {
         path: "settings",