+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 2022-01-13 (3.7.0)
+
+* Fix upload widget toggles on interaction with projects list
+* Add `WITH_PROJECTS` config parameter
+* Add `WITH_REMOTE_UPLOAD` config parameter
+* Add `WITH_START_PRINT_AFTER_UPLOAD` config parameter
+* Add `WITH_CONNECTION` config parameter
+* Add `WITH_SERIAL` config parameter
+* Fix sidebar width
+* Replace PNG icons with SVG
+
 ## 2022-01-13 (3.6.0)
 
 * Add stop/resume print button